feat(generation): all generation works
The script generation_all.py do the PSTR generation of outputs. From these outputs we need to get the automatic crops from them and also we will need the features of the crop from the annotations. Code base as been simplified and refactorised.

no much comment sorry, do not have the time to document, hope the code is self documented enough (:

!! Automatic crops with current settings takes 7G of VRAM and 50mn of processing (A4000). The generation